[fix] 코드 리뷰 피드백 반영(운채님) - 서비스 측과 리포지토리 측의 명확한 책임 분리

+++ src/main/java/com/example/aniwhere/domain/episodeReviews/dto/EpisodeReviewResponse.java
@@ -4,19 +4,18 @@ import com.querydsl.core.annotations.QueryProjection;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 
 @Getter
 @Setter
 public class EpisodeReviewResponse {
 
 	private final Long id;
-	private final BigDecimal rating;
+	private final Double rating;
 	private final String content;
 	private final Long userId;
 
 	@QueryProjection
-	public EpisodeReviewResponse(Long id, BigDecimal rating, String content, Long userId) {
+	public EpisodeReviewResponse(Long id, Double rating, String content, Long userId) {
 		this.id = id;
         this.rating = rating;
         this.content = content;
